Extract CurrentLocationOverlay to util.map package
